+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -745,7 +745,7 @@ class RanaBrowser(QWidget):
     create_model_selected = pyqtSignal(dict, dict)
     create_model_selected_with_revision = pyqtSignal(dict, dict, int)
     delete_model_selected = pyqtSignal(dict, int)
-    open_schematisation_selected_with_revision = pyqtSignal(dict, dict)
+    open_schematisation_selected_with_revision = pyqtSignal(dict, dict, dict)
     delete_file_selected = pyqtSignal(dict, dict)
     rename_file_selected = pyqtSignal(dict, dict, str)
     create_folder_selected = pyqtSignal(dict, dict, str)
@@ -754,6 +754,7 @@ class RanaBrowser(QWidget):
     request_monitoring_project_jobs = pyqtSignal(str)
     project_jobs_added = pyqtSignal(list)
     project_job_updated = pyqtSignal(dict)
+    update_avatar_cache = pyqtSignal(list)
     view_file_after_open = pyqtSignal(dict)
 
     def __init__(self, communication: UICommunication):
@@ -785,7 +786,6 @@ class RanaBrowser(QWidget):
             avatar_cache=self.avatar_cache,
             parent=self,
         )
-        self.avatar_cache.update_users_in_thread(self.projects_browser.users)
         file_signals = FileActionSignals()
         self.files_browser = FilesBrowser(
             communication=self.communication, file_signals=file_signals, parent=self
@@ -868,9 +868,7 @@ class RanaBrowser(QWidget):
         # Connect avatar_cache
         # Note that avatar_cache is only linked to the projects_browser because for now
         # all widgets that use avatars are loaded after the projects browser is initialized
-        self.projects_browser.users_refreshed.connect(
-            self.avatar_cache.update_users_in_thread
-        )
+        self.projects_browser.users_refreshed.connect(self.update_avatar_cache.emit)
         self.avatar_cache.avatar_changed.connect(self.projects_browser.update_avatar)
 
         # Disable/enable widgets
@@ -886,6 +884,7 @@ class RanaBrowser(QWidget):
         )
         self.project_jobs_added.connect(self.processes_browser.add_items)
         self.project_job_updated.connect(self.processes_browser.update_job_state)
+
         # Connect refresh buttons
         self.projects_browser.refresh_btn.clicked.connect(self.refresh_projects_browser)
         refresh_btn.clicked.connect(self.refresh_project_widget)
@@ -997,7 +996,10 @@ class RanaBrowser(QWidget):
         )
         # Load specific revision of schematisation
         self.revisions_view.open_schematisation_revision_in_qgis_requested.connect(
-            self.open_schematisation_selected_with_revision
+            lambda schematisation,
+            revision: self.open_schematisation_selected_with_revision.emit(
+                self.project, schematisation, revision
+            )
         )
         # Update breadcrumbs when file browser path changes
         self.projects_browser.project_selected.connect(