+++ HISTORY.rst
@@ -2,14 +2,33 @@ History
 =======
 
 
-1.2.7 (unreleased)
+1.2.9 (unreleased)
+------------------
+
+- Handle uncaught errors via custom excepthook (nens/rana-qgis-plugin#254)
+
+
+1.2.8 (2026-03-13)
+------------------
+
+- Proper loading of downloaded geopackages (nens/rana#3781)
+
+
+1.2.7 (2026-03-13)
 ------------------
 
 - Automatically open file detail when opening file in qgis via context menu (nens/rana-qgis-plugin#275)
 - Prevent deleting results replacing an existing revision (nens/rana-qgis-plugin#259)
 - Create model when creating new schematisation (nens/rana#3533)
 - Add cancel button for running simulations (nens/rana-qgis-plugin#235)
-- Handle uncaught errors via pyqt signals (nens/rana-qgis-plugin#254)
+- Remove confirmation dialog on opening downloaded schematisations (nens/rana-qgis-plugin#289)
+- Take buttons in the file view out of the scroll area (nens/rana#3630)
+- Make creating avatars safe for users without given and/or family name
+- Download results folder "<simulation name> (<simulation id>)" (nens/rana#3678)
+- Match update interval of project monitoring to update interval in rana
+- Enable results downloading for results without lizard data and results not linked to a simulation (nens/rana#3710)
+- Organize opened layers under Rana projects (nens/rana-qgis#57)
+- Basic setup for end-to-end testing (nens/rana-plugin#3196)
 
 
 1.2.6 (2026-02-19)